improve

+++ backup.py
@@ -4,7 +4,8 @@
 
 Usage:
   backup.py create [--db=<db.txt>] [--fs=<local>]
-  backup.py check (local | external | rsync | <path>) [--db=<db.txt>]
+  backup.py update [--db=<db.txt>] [--fs=<local>]
+  backup.py check (local | external | rsync | <path>) [--db=<db.txt>] [--checksum]
   backup.py (-h | --help)
   backup.py [--verbose | -v | -vv | -vvv]
   
@@ -23,7 +24,7 @@ import traceback
 import logging
 
 def init_logging():
-    LOG_LEVEL = logging.CRITICAL
+    LOG_LEVEL = logging.INFO
     LOGFORMAT = "  %(log_color)s%(levelname)-8s%(reset)s | %(log_color)s%(message)s%(reset)s"
     from colorlog import ColoredFormatter
     logging.root.setLevel(LOG_LEVEL)
@@ -51,18 +52,9 @@ GOOD_FILES = "good.txt"
 STARTUP_CLEANING = MISSING_FILES, DIFFERENT_FILES, NEW_FILES, GOOD_FILES
 
 def progress(current, total):
-    sys.stdout.write("\r%.2f%%" % (current/total*100))
-    sys.stdout.flush()
-
-class Checksym(Enum):
-    MD5_FILE = 1
-
-CHKSUM_ALGO = True # or None
-    
+    print("\r%.2f%%" % (current/total*100), end="")
+        
 def checksum(fname):
-    if not CHKSUM_ALGO:
-        return "0"
-    
     hashval = hashlib.md5()
         
     with open(fname, "rb") as f:
@@ -84,7 +76,7 @@ def browse_db(db_fname):
 
             yield relpath, info
 
-def browse_filesystem(fs_dir):
+def browse_filesystem(fs_dir, do_checksum):
     for dirpath, dirnames, files in os.walk(fs_dir):
         
         for ign in TO_IGNORE:
@@ -102,60 +94,80 @@ def browse_filesystem(fs_dir):
             st = os.stat(fullpath)
             
             info = {
-                "filesum" : checksum(fullpath),
+                "md5sum" : checksum(fullpath) if do_checksum else "",
                 "size" : str(st.st_size)
-                #, "mtime" :  str(st.st_mtime)
                 }
 
             yield fullpath, relpath, info
         yield None
 
-def print_filesystem(fs_dir, out_f=sys.stdout):
-    for a_file in browse_filesystem(fs_dir):
+def print_filesystem(fs_dir, out_f=sys.stdout, do_checksum=True):
+    for a_file in browse_filesystem(fs_dir, do_checksum):
         if a_file is None:
             if out_f == sys.stdout:
                 print("---")
             continue
-        fullpath, relpath, info = a_file
-        print("{} -> {}".format(fullpath.replace(fs_dir+"/", ""),
-                                ", ".join([ "{}: {}".format(k, v) for k,v in info.items()])),
-              file=out_f)
+        
+        print_a_file(a_file, out_f)
+
+def print_a_file(a_file, out_f):
+    relpath, info = a_file[-2:]
+    print("{} -> {}".format(relpath,
+                            ", ".join([ "{}: {}".format(k, v) for k,v in info.items()])),
+          file=out_f)
+
 
 def create_database(db_file, fs_dir):
+    if os.path.exists(db_file):
+        log.critical("Database file '{}' already exists. Delete it first to (re)create the database.".format(db_file))
+        return
+    
     with open(db_file, "w+") as db_f:
         print_filesystem(fs_dir, out_f=db_f)
 
-def compare_entries(fs_dir, fs_entry, db_entry):
+def compare_entries(fs_dir, fs_entry, db_entry, do_checksum):
     fs_fullpath, fs_relpath, fs_info = fs_entry
     db_relpath, db_info = db_entry
 
     db_fullpath = "{}/{}".format(fs_dir, db_relpath)
 
     if fs_fullpath != db_fullpath:
-        log.error("DB/FS mismatch ... {} != {}".format(db_relpath, fs_relpath))
+        missing_on_fs = db_fullpath > fs_fullpath
+            
+        log.warn("{} # {}".format(fs_relpath, "missing" if missing_on_fs else "new"))
         
-        return db_fullpath > fs_fullpath, None
+        return missing_on_fs, None
 
     errors = {}
     
     for key, fs_val in fs_info.items():
+        if key == "md5sum" and not do_checksum:
+            # ignore checksum
+            continue
+        
         db_val = db_info[key]
+        
         if db_val == fs_val:
+            # value is correct
             continue
     
         errors[key] = (db_val, fs_val)
 
     if errors:
-        log.warning("{} on {}".format(db_relpath, ", ".join(errors.keys())))
+        log.warning("{} # different {}".format(db_relpath, ", ".join(errors.keys())))
     
     return None, errors
 
-def compare_fs_db(fs_dir, db_file):
+def compare_fs_db(fs_dir, db_file, do_checksum):
+    if not os.path.exists(fs_dir):
+        log.critical("Path '{}' does not exist.".format(fs_dir))
+        return
+    
     total_len = db_length(db_file)
     
     count = 0
     db = browse_db(db_file)
-    fs = browse_filesystem(fs_dir)
+    fs = browse_filesystem(fs_dir, do_checksum)
     ret = None
 
     try:
@@ -163,6 +175,8 @@ def compare_fs_db(fs_dir, db_file):
         missing_f = open(MISSING_FILES, "w+")
         different_f = open(DIFFERENT_FILES, "w+")
         good_f = open(GOOD_FILES, "w+")
+
+        good, different, missing, new = 0, 0, 0, 0
         
         while True:
             if ret is None:
@@ -181,36 +195,47 @@ def compare_fs_db(fs_dir, db_file):
                 # new directory
                 continue 
             
-            count += 1
+            
             # returns None if could compare,
             #      or db_fullpath > fs_fullpath
-            ret, diff = compare_entries(fs_dir, fs_entry, db_entry)
+            ret, diff = compare_entries(fs_dir, fs_entry, db_entry, do_checksum)
             
             fs_fullpath, fs_relpath, fs_info = fs_entry
             db_relpath, db_info = db_entry
 
             progress(count, total_len)
+            count += 1
             
             if ret is None:
                 if not diff:
                     # files are identical
                     print(db_relpath, file=good_f)
+                    good += 1
                 else:
                     # there are some differences, in diff dict
                     print("{}: {}".format(db_relpath, ", ".join(diff.keys())),
                           file=different_f)
+                    different += 1
             else:
                 # one of the files is missing
                 if ret:
                     # on the database
                     print(fs_fullpath, file=new_f)
+                    new += 1
                 else:
                     # on the filesystem
                     print(db_relpath, file=missing_f)
+                    missing += 1
                 
                 
     except StopIteration:
-        log.info("Done, {} files compared.".format(count))
+        print("")
+        
+        log.critical("Done, {} files compared.".format(count))
+        log.error("Good files: {}".format(good))
+        log.error("Missing files: {}".format(missing))
+        log.error("Different files: {}".format(different))
+        log.error("New files: {}".format(new))
         
     finally:
         if missing_f:
@@ -219,6 +244,95 @@ def compare_fs_db(fs_dir, db_file):
             different_f.close()
         if new_f:
             new_f.close()
+
+
+def update_database(db_file, fs_dir, do_checksum):
+    total_len = db_length(db_file)
+    
+    count = 0
+    db = browse_db(db_file)
+    fs = browse_filesystem(fs_dir, do_checksum)
+    ret = None
+    updated, new, missing, untouched = 0, 0, 0, 0
+
+    tmp_db_file = "{}.tmp".format(db_file)
+    
+    try:
+        tmp_db_f = open(tmp_db_file, "w+")
+        
+        while True:
+            if ret is None:
+                fs_entry = next(fs)
+                
+                if fs_entry is not None:
+                    # not a new directory
+                    db_entry = next(db)
+            
+            elif ret: # is True
+                fs_entry = next(fs)
+            else: # ret is False
+                db_entry = next(db)
+
+            if fs_entry is None:
+                # new directory
+                continue 
+            
+            
+            # returns None if could compare,
+            #      or db_fullpath > fs_fullpath
+            ret, diff = compare_entries(fs_dir, fs_entry, db_entry, do_checksum)
+            
+            fs_fullpath, fs_relpath, fs_info = fs_entry
+            db_relpath, db_info = db_entry
+
+            progress(count, total_len)
+            count += 1
+
+            a_file = fs_entry
+            if ret is None:
+                if not diff:
+                    # files are identical
+                    untouched += 1
+                    a_file = db_entry
+                else:
+                    # there are some differences, in diff dict
+                    updated += 1
+                    if not do_checksum:
+                        # force compute the checksum if disabled
+                        ds_entry["md5sum"] = checksum(fullpath)
+            else:
+                # the file is missing
+                if ret:
+                    # on the database
+                    new += 1
+                else:
+                    # on the filesystem
+                    missing += 1
+                    continue # skip this entry
+                
+            print_a_file(a_file, tmp_db_f)
+            
+    except StopIteration:
+        print("")
+
+        log.critical("Database updated.")
+        log.error("{} entries untouched".format(untouched))
+        log.error("{} entries removed".format(missing))
+        log.error("{} entries updated".format(updated))
+        log.error("{} entries added".format(new))
+
+        tmp_db_f.close()
+        tmp_db_f = None # don't close it twice
+
+        
+        #try: os.remove(db_file)
+        #except OSError: pass
+
+        #os.rename(tmp_db_file, db_file)
+        
+    finally:
+        if tmp_db_f:
+            tmp_db_f.close()
             
 class Done(Exception): pass
     
@@ -226,10 +340,8 @@ DEFAULT_DB_FILE = "db.txt"
 DEFAULT_FS_DIR = "/var/webalbums"
 NOP = False
 
-if __name__ == '__main__':
+def main(args):
     try:
-        args = docopt(__doc__, version='backup.py 0.9')
-
         db_file = args["--db"]
         if not db_file:
             db_file = DEFAULT_DB_FILE
@@ -237,14 +349,25 @@ if __name__ == '__main__':
         fs_dir = args["--fs"]
         if not fs_dir:
             fs_dir = DEFAULT_FS_DIR
+
+        do_checksum = args["--checksum"]
+            
+        #################
             
         if args["create"]:
-            print("create database {} from {}".format(db_file, fs_dir))
+            print("Create database {} from {}".format(db_file, fs_dir))
             
             if NOP: raise Done()
 
             create_database(db_file, fs_dir)
-        
+            
+        elif args["update"]:
+            print("Update database {} from {}".format(db_file, fs_dir))
+            
+            if NOP: raise Done()
+
+            update_database(db_file, fs_dir, do_checksum)
+            
         elif args["check"]:        
             if args["local"]:
                 fs_file = "/var/webalbums"
@@ -255,7 +378,8 @@ if __name__ == '__main__':
             else:
                 fs_file = args["<path>"]
             
-            print("check {} against {}".format(fs_file, db_file))
+            print("Check {} against {} ({})".format(fs_file, db_file,
+                                                    "checksum" if do_checksum else "fast"))
             
             if NOP: raise Done()
 
@@ -264,7 +388,7 @@ if __name__ == '__main__':
                     os.remove(to_remove)
                 except OSError: pass
             
-            compare_fs_db(fs_file, db_file)
+            compare_fs_db(fs_file, db_file, do_checksum)
             
         else:
             print(__doc__)
@@ -280,3 +404,6 @@ if __name__ == '__main__':
             print("  "+"\n  ".join("{:10s}->  {}".format(k, v) for k, v in args.items() if v))
             print("---")
             print("  "+"\n  ".join("{:10s}->  {}".format(k, v) for k, v in args.items() if not v))
+            
+if __name__ == '__main__':
+    main(docopt(__doc__, version='backup.py 0.9'))
